ioctlsort can run on the host.
Upstream has changed how all this stuff works in support of being able
to say:

  ioctl(1, TCGETS, {B38400 opost isig icanon echo ...}) = 0

instead of the usual:

  ioctl(1, SNDCTL_TMR_TIMEBASE or SNDRV_TIMER_IOCTL_NEXT_DEVICE or TCGETS, {B38400 opost isig icanon echo ...}) = 0

The problem for us is that they now do their ioctlsort stuff at build
time, rather than have the maintainer do it once. In case we have to
follow suit, this at least removes the need to run code on the target!

Change-Id: I94a9a07c8969d9b228846faa491fdef02e545206